Rename MCP runtime module
Rename mcp_connection.rs to runtime.rs now that the module only contains runtime and sandbox shared types plus the shared duration metric helper.

Update imports and exports to use the new runtime module name.
