+++ codex-rs/codex-mcp/src/client.rs
@@ -20,8 +20,8 @@ use crate::apps::write_cached_codex_apps_tools_if_needed;
 use crate::elicitation::ElicitationRequestManager;
 use crate::mcp::CODEX_APPS_MCP_SERVER_NAME;
 use crate::mcp::ToolPluginProvenance;
-use crate::mcp_connection::McpRuntimeEnvironment;
-use crate::mcp_connection::emit_duration;
+use crate::runtime::McpRuntimeEnvironment;
+use crate::runtime::emit_duration;
 use crate::tools::ToolFilter;
 use crate::tools::ToolInfo;
 use crate::tools::filter_tools;
