+++ codex-rs/codex-mcp/src/manager.rs
@@ -18,8 +18,8 @@ use crate::client::list_tools_for_client_uncached;
 use crate::elicitation::ElicitationRequestManager;
 use crate::mcp::CODEX_APPS_MCP_SERVER_NAME;
 use crate::mcp::ToolPluginProvenance;
-use crate::mcp_connection::McpRuntimeEnvironment;
-use crate::mcp_connection::emit_duration;
+use crate::runtime::McpRuntimeEnvironment;
+use crate::runtime::emit_duration;
 use crate::tools::ToolInfo;
 use crate::tools::filter_tools;
 use crate::tools::qualify_tools;
