+++ codex-rs/codex-mcp/src/mcp/mod.rs
@@ -36,7 +36,7 @@ use serde_json::Value;
 
 use crate::apps::codex_apps_tools_cache_key;
 use crate::manager::McpConnectionManager;
-use crate::mcp_connection::McpRuntimeEnvironment;
+use crate::runtime::McpRuntimeEnvironment;
 
 pub const CODEX_APPS_MCP_SERVER_NAME: &str = "codex_apps";
 const MCP_TOOL_NAME_PREFIX: &str = "mcp";
